Improve dependency handling, create proper internal shlib
